Split EngineTracer into two independent classes
