+++ src/Engine/Components/Trace/CMakeLists.txt
@@ -2,12 +2,14 @@ cmake_minimum_required(VERSION 3.20.4 FATAL_ERROR)
 
 set(ENGINE_COMPONENTS_TRACE_SOURCES
         EngineTraceComponent.cpp
-        EngineTracer.cpp)
+        EngineTracePlayer.cpp
+        EngineTraceRecorder.cpp)
 
 set(ENGINE_COMPONENTS_TRACE_HEADERS
         EngineTraceComponent.h
         EngineTracePlaybackFlags.h
-        EngineTracer.h)
+        EngineTracePlayer.h
+        EngineTraceRecorder.h)
 
 add_library(engine_components_trace STATIC ${ENGINE_COMPONENTS_TRACE_SOURCES} ${ENGINE_COMPONENTS_TRACE_HEADERS})
 target_check_style(engine_components_trace)
