--- src/Engine/Components/Trace/EngineTracer.cpp
@@ -1,150 +1,0 @@
-#include "EngineTracer.h"
-
-#include <cassert>
-#include <utility>
-#include <filesystem>
-
-#include "Application/GameKeyboardController.h" // TODO(captainurist): Engine -> Application dependency
-
-#include "Engine/Engine.h"
-#include "Engine/Components/Control/EngineController.h"
-#include "Engine/Components/Control/EngineControlComponent.h"
-#include "Engine/Components/Deterministic/EngineDeterministicComponent.h"
-
-#include "Library/Trace/PaintEvent.h"
-#include "Library/Application/PlatformApplication.h"
-#include "Library/Random/Random.h"
-
-#include "Utility/ScopeGuard.h"
-#include "Utility/Exception.h"
-
-#include "EngineTraceComponent.h"
-
-EngineTracer::EngineTracer(Options options): _options(options) {}
-
-EngineTracer::~EngineTracer() {
-    assert(!application()); // We're uninstalled.
-}
-
-void EngineTracer::startTraceRecording(const std::string &savePath, const std::string &tracePath) {
-    assert(_options & ENABLE_RECORDING);
-    assert(_state == CHILLING);
-
-    _state = RECORDING;
-    _saveFilePath = savePath;
-    _traceFilePath = tracePath;
-
-    // TODO(captainurist): this is actually unsafe, we need to cancel running routines in dtor if there are any.
-    _controlComponent->runControlRoutine([this, savePath] (EngineController *game) {
-        game->resizeWindow(640, 480);
-        game->tick();
-
-        _oldFpsLimit = engine->config->graphics.FPSLimit.Get();
-        engine->config->graphics.FPSLimit.Set(1000); // Load game real quick!
-        game->saveGame(savePath);
-        _deterministicComponent->enterDeterministicMode();
-        game->loadGame(savePath);
-        _deterministicComponent->resetDeterministicState();
-        _keyboardController->reset(); // Reset all pressed buttons.
-
-        _traceComponent->start();
-        engine->config->graphics.FPSLimit.Set(EngineDeterministicComponent::TARGET_FPS); // But don't turn the party into a wall-running doomguy.
-
-        logger->Info("Tracing started.");
-    });
-}
-
-void EngineTracer::finishTraceRecording() {
-    assert(_options & ENABLE_RECORDING);
-    assert(_state == RECORDING);
-
-    // The code below needs to go into a control routine because it should be run AFTER the previous control routine
-    // is finished, and it could take a while for it to finish.
-    _controlComponent->runControlRoutine([this, tracePath = _traceFilePath, savePath = _saveFilePath] (EngineController *game) {
-        assert(_traceComponent->isTracing()); // Just a sanity check.
-        _deterministicComponent->leaveDeterministicMode();
-        engine->config->graphics.FPSLimit.Set(_oldFpsLimit);
-
-        EventTrace trace = _traceComponent->finish();
-        trace.header.saveFileSize = std::filesystem::file_size(savePath); // This function can throw.
-        EventTrace::saveToFile(tracePath, trace);
-
-        logger->Info("Trace saved to {} and {}",
-                     absolute(std::filesystem::path(tracePath)).generic_string(),
-                     absolute(std::filesystem::path(savePath)).generic_string());
-    });
-
-    _state = CHILLING;
-    _traceFilePath.clear();
-    _saveFilePath.clear();
-}
-
-void EngineTracer::playTrace(EngineController *game, const std::string &savePath, const std::string &tracePath,
-                             EngineTracePlaybackFlags flags, std::function<void()> postLoadCallback) {
-    assert(_options & ENABLE_PLAYBACK);
-    assert(_state == CHILLING);
-
-    _state = PLAYING;
-    MM_AT_SCOPE_EXIT(_state = CHILLING);
-
-    EventTrace trace = EventTrace::loadFromFile(tracePath, application()->window());
-
-    int saveFileSize = std::filesystem::file_size(savePath);
-    if (trace.header.saveFileSize != -1 && trace.header.saveFileSize != saveFileSize) {
-        throw Exception("Trace '{}' expected a savegame of size {} bytes, but the size of '{}' is {} bytes",
-                        tracePath, trace.header.saveFileSize, savePath, saveFileSize);
-    }
-
-    game->resizeWindow(640, 480);
-    game->tick();
-
-    _deterministicComponent->enterDeterministicMode();
-    MM_AT_SCOPE_EXIT(_deterministicComponent->leaveDeterministicMode());
-
-    game->loadGame(savePath);
-    _deterministicComponent->resetDeterministicState();
-    _keyboardController->reset(); // Reset all pressed buttons.
-
-    if (postLoadCallback)
-        postLoadCallback();
-
-    for (std::unique_ptr<PlatformEvent> &event : trace.events) {
-        if (event->type == EVENT_PAINT) {
-            game->tick(1);
-
-            const PaintEvent *paintEvent = static_cast<const PaintEvent *>(event.get());
-
-            int64_t tickCount = application()->platform()->tickCount();
-            if (!(flags & TRACE_PLAYBACK_SKIP_TIME_CHECKS) && tickCount != paintEvent->tickCount) {
-                throw Exception("Tick count desynchronized when playing back trace '{}': expected {}, got {}",
-                                tracePath, paintEvent->tickCount, tickCount);
-            }
-
-            int randomState = grng->Random(1024);
-            if (!(flags & TRACE_PLAYBACK_SKIP_RANDOM_CHECKS) && randomState != paintEvent->randomState) {
-                throw Exception("Random state desynchronized when playing back trace '{}' at {}ms: expected {}, got {}",
-                                tracePath, tickCount, paintEvent->randomState, randomState);
-            }
-        } else {
-            game->postEvent(std::move(event));
-        }
-    }
-}
-
-void EngineTracer::installNotify() {
-    _controlComponent = application()->get<EngineControlComponent>();
-    _traceComponent = application()->get<EngineTraceComponent>();
-    _deterministicComponent = application()->get<EngineDeterministicComponent>();
-    _keyboardController = application()->get<GameKeyboardController>();
-    assert((_options & ENABLE_RECORDING) ? _controlComponent && _deterministicComponent && _traceComponent && _keyboardController : true);
-    assert((_options & ENABLE_PLAYBACK) ? _controlComponent && _deterministicComponent && _keyboardController : true);
-}
-
-void EngineTracer::removeNotify() {
-    _controlComponent = nullptr;
-    _traceComponent = nullptr;
-    _deterministicComponent = nullptr;
-    _keyboardController = nullptr;
-}
-
-
